Added class Helpers
