+++ tests/GitPhp/bootstrap.php
@@ -1,5 +1,8 @@
 <?php
 require __DIR__ . '/../../vendor/nette/tester/Tester/bootstrap.php';
+require __DIR__ . '/../../src/exceptions.php';
+require __DIR__ . '/../../src/Helpers.php';
+require __DIR__ . '/../../src/GitRepository.php';
 
 // create temporary directory
 define('TEMP_DIR', __DIR__ . '/../tmp/' . (isset($_SERVER['argv']) ? md5(serialize($_SERVER['argv'])) : getmypid()));
